@@ -1,5 +1,3 @@
-/* global document:true */
-
 import {connect} from 'react-redux';
 import {any} from 'ramda';
 import {Component} from 'react';
@@ -13,7 +11,7 @@ class DocumentTitle extends Component {
         };
     }
 
-    componentWillReceiveProps(props) {
+    UNSAFE_componentWillReceiveProps(props) {
         if (any(r => r.status === 'loading', props.requestQueue)) {
             document.title = 'Updating...';
         } else {
