@@ -1,4 +1,3 @@
-/* global fetch:true, Promise:true, document:true */
 import {
     adjust,
     any,
@@ -429,9 +428,8 @@ export function notifyObservers(payload) {
             );
         }
 
-        /* eslint-disable consistent-return */
+        /* eslint-disable-next-line consistent-return */
         return Promise.all(promises);
-        /* eslint-enable consistent-return */
     };
 }
 
@@ -645,13 +643,13 @@ function updateOutput(
                 return;
             }
 
-            /* eslint-disable no-console */
+            /* eslint-disable-next-line no-console */
             console.error(
                 `The following error occurred while executing ${clientside_function.namespace}.${clientside_function.function_name} ` +
                     `in order to update component "${payload.output}" ⋁⋁⋁`
             );
+            /* eslint-disable-next-line no-console */
             console.error(e);
-            /* eslint-enable no-console */
 
             /*
              * Update the request queue by treating an unsuccessful clientside
@@ -665,7 +663,7 @@ function updateOutput(
 
         // Returning promises isn't support atm
         if (type(returnValue) === 'Promise') {
-            /* eslint-disable no-console */
+            /* eslint-disable-next-line no-console */
             console.error(
                 'The clientside function ' +
                     `${clientside_function.namespace}.${clientside_function.function_name} ` +
@@ -673,7 +671,6 @@ function updateOutput(
                     'supported in Dash clientside right now, but may be in the ' +
                     'future.'
             );
-            /* eslint-enable no-console */
             updateRequestQueue(true, STATUS.CLIENTSIDE_ERROR);
             return;
         }
@@ -736,12 +733,10 @@ function updateOutput(
         hooks.request_pre(payload);
     }
 
-    /* eslint-disable consistent-return */
+    /* eslint-disable-next-line consistent-return */
     return fetch(
         `${urlBase(config)}_dash-update-component`,
         mergeDeepRight(config.fetch, {
-            /* eslint-enable consistent-return */
-
             method: 'POST',
             headers: getCSRFHeader(),
             body: JSON.stringify(payload),
