@@ -1,10 +1,8 @@
-'use strict';
-
 export default {
     resolve: component => {
         const {type, namespace} = component;
 
-        const ns = window[namespace]; /* global window: true */
+        const ns = window[namespace];
 
         if (ns) {
             if (ns[type]) {
